+++ sdk/python/pulumi_okta/app/group_assignment.py
@@ -27,28 +27,6 @@ class GroupAssignment(pulumi.CustomResource):
 
         This resource allows you to create an App Group assignment.
 
-        __When using this resource, make sure to add the following `lifefycle` argument to the application resource you are assigning to:__
-
-        ```python
-        import pulumi
-        ```
-
-        ## Example Usage
-
-        ```python
-        import pulumi
-        import pulumi_okta as okta
-
-        example = okta.app.GroupAssignment("example",
-            app_id="<app id>",
-            group_id="<group id>",
-            profile=\"\"\"{
-          "<app_profile_field>": "<value>"
-        }
-
-        \"\"\")
-        ```
-
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] app_id: The ID of the application to assign a group to.